workaround to fix the home link in navbar
set the home and gotta eat links that redirect to index.php to actually
redirect to index.php when on another page besides index.php. only
catch: all our files have to reside in ../ajax_recipe/ for this to work.
if this root directory is ever changed, the code i created has to be
changed as well; we could do this with global constants or whatever if
you fee adventurous.